Change: don't build Jekyll on docker run, but on docker build
The initial idea was the run Jekyll on run, so it could fetch
dynamic information, like the top left banner with Download
information. This complicates things a lot, so for the initial
version we switch to a multi-stage Dockerfile where Jekyll is
build on docker build, and we run a full static version.