Fix: for yet-unknown-reasons 'bundle install' no longer installs http_parser.rb and safe_yaml
A fix seems to be to install it via gem. As I don't understand
Ruby (and gem and bundle), I am guessing we are missing something
here, but .. it works again. As Ruby is never run in production,
at this point, I also don't care enough to find out what exactly
is happening here.