Add: downloads pages, and links from header/index to them
When the docker is build, information is downloaded (via HTTP) from
finger.openttd.org and binaries.openttd.org to figure out what
the available downloads are, and those are added to the collection.
Jekyll after that makes this into static files.

Yes, this does mean we have to rebuild this container every time
we make a new release. But those events are rare.