Fix: limit "openttd-nightlies" to 14
It currently was generating the HTML of the last 10 months (and
increasing). This was giving more and more issues due to the
instability of the current CDN in use.

As there is no need to generate HTML for all the nightlies anyway,
limit it to the last 14 is a good idea anyway. For those that want
to access older nightlies, they can use the archive directly.