Update: bump all Python dependencies to latest version
While at it, switch to Python 3.8, and to the "slim" version.
For the "alpine" version pip cannot download the precompiled
binaries, which means that it needs gcc and friends to build
them. This is pretty much unwanted.
Switching to "slim" means pip can download the precompiled
binaries, resolving the issue.