Add: switch to cdn.openttd.org for all downloads
Here also is a file called "latest.yaml", which greatly simplifies
the "fetch-downloads.py". Additionally, as cdn.openttd.org is
browseable, we only need to serve "latest" (and no longer older
releases).

While at it, refactored fetch_downloads a bit to be a Python
module and run flake8/black on it.