Remove: we no longer use a mirror network that depends on donations (#164)
We are now using AWS CDN, and most likely will for the foreseeable
future. Maintaining your own CDN is a lot of work, with a lot of
quirks, and a lot of monitoring. Using an existing CDN, although
a bit more pricy (donated vs paid) does relieve a lot of SysOp
burden.

We do like to thank all the people who donated a mirror over the
years. It is much appreciated!