Add: port the security patches we have for CVEs (#165)
Instead of running on its own site, it now integrates with the
main website. This avoids copying the Jekyll website and maintaining
two instances of it. The main website and the CVE listing are similar
enough that this shouldn't be an issue.

The data is imported from the current security listing we have,
including the patches available. It is imported as-is without
modification.