Add: rework our "Contact" page
It no longer shows email-addresses of Developers, but instead
gives a whole bunch of places people could go to for the contact
they are looking for.

This is mostly triggered by the increased amount of questions
we get at info@, which is not a place you will receive an answer
from.