Fix: [devcontainer] optimize launch speed by using a pre-build Python
If we use "features", the devcontainer will build (from source!)
Python on-demand. This takes ~3 minutes. We can skip this by copying
the Python interpreter from the pre-build Python image, while building
our devcontainer.

This is basically a quirk of the fact we are a hybrid repo: both
Ruby and Python are used. This is somewhat unusual.